requirements.txt
matplotlib==3.4.3
numpy==1.21.0
Pillow==8.3.2
PyQt5==5.15.6